Add UserProfile class for DTO to model mappings
This commit introduces the `UserProfile` class in the
`Aptiverse.Application.Users.Mapping` namespace, which
utilizes AutoMapper to facilitate mappings between
`UserDto` and `User`, `UserType` and `UserTypeDto`,
as well as `UserCategory` and `UserCategoryDto`.
The `Id` property of the `User` is configured to
always map to `0` when creating a `User` from a
`UserDto`, with support for reverse mapping.